Remove prefs only when uninstalling

+++ bootstrap.js
@@ -53,7 +53,8 @@ function processWindow(win) {
 
 function install(data, reason) {}
 function uninstall(data, reason) {
-	prefBranch.deleteBranch("");
+	if (reason == ADDON_UNINSTALL)
+		prefBranch.deleteBranch("");
 }
 function startup(data, reason) {
 	let ss = Cc["@mozilla.org/content/style-sheet-service;1"].getService(Ci.nsIStyleSheetService);
